naive hypervisor occupancy check
